tests

+++ build.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 echo "Building $0/$1"
-rm $1.o
-rm $1.ko
+if [ -f $1.o ] 
+then 
+    rm $1.o 
+fi
+if [ -f $1.ko ] 
+then 
+    rm $1.ko 
+fi
 echo "obj-m += $1.o" > Makefile
 make -C ../linux-3.3-rc2/ SUBDIRS=$PWD modules 
 rmmod $1.
